requirements.txt
transformers
torch
requests 
beautifulsoup4 
pandas 
numpy
vaderSentiment
selenium
webdriver-manager
google-api-python-client
google-auth 
google-auth-oauthlib 
google-auth-httplib2
streamlit
streamlit-auth0-component
PyPDF
docx2txt